requirements.txt
anyio==2.0.2
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
Babel==2.9.0
backcall==0.2.0
bleach==3.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cmsplugin-filer==1.1.3
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.2
django-appconf==1.0.4
django-classy-tags==2.0.0
django-cms==3.8.0
django-crispy-forms==1.9.2
django-filer==2.0.2
django-formtools==2.2
django-jalali==4.0.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-polymorphic==3.0.0
django-ranged-response==0.2.0
django-recaptcha==2.0.6
django-reversion==3.0.8
django-sekizai==2.0.0
django-simple-captcha==0.5.13
django-taggit==1.3.0
django-treebeard==4.3.1
djangocms-admin-style==2.0.1
djangocms-attributes-field==2.0.0
easy-thumbnails==2.7
entrypoints==0.3
filelock==3.0.12
html5lib==1.1
idna==2.10
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.5.4
jdatetime==3.6.2
jedi==0.18.0
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyter-server==1.2.2
jupyterlab==3.0.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
notebook==6.2.0
numpy==1.19.5
packaging==20.8
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
persian==0.4.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.9.0
prompt-toolkit==3.0.13
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.4
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pywin32==300
pywinpty==0.5.7
pyzmq==21.0.1
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
sqlparse==0.3.1
terminado==0.9.2
testpath==0.4.4
toml==0.10.1
tornado==6.1
traitlets==5.0.5
Unidecode==1.1.1
urllib3==1.26.2
virtualenv==20.1.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
